+++ src/TRegx/CleanRegex/Match/Details/Group/MatchGroupDetails.php
@@ -34,5 +34,7 @@ interface MatchGroupDetails extends Optional
 
     public function replace(string $replacement): string;
 
+    public function subject(): string;
+
     public function all(): array;
 }